requirements.txt
azure-ai-ml ~= 1.21.1
pydantic ~= 2.9.2
typing_extensions ~= 4.12.2
